letsencrypt-cert-autorenewal/requirements.txt
# Azure Key Vault Certificate Renewal - Python Dependencies

# Azure SDK for Key Vault operations
azure-identity>=1.15.0
azure-keyvault-certificates>=4.8.0
azure-core>=1.29.0

# AWS SDK for OIDC authentication (AssumeRoleWithWebIdentity)
boto3>=1.34.0
botocore>=1.34.0

# HTTP client for fetching GitHub OIDC token
requests>=2.31.0

# Certificate parsing (for extracting domains from X.509 certificates)
cryptography>=41.0.0

# Certbot and DNS plugins
certbot>=2.8.0
certbot-dns-route53>=2.8.0
certbot-dns-azure>=2.0.0

# Pin azure-mgmt-dns to 8.2.0 due to breaking changes in 9.0.0
# See: https://github.com/certbot/certbot/issues/10367
azure-mgmt-dns==8.2.0

# Configuration file parsing
PyYAML>=6.0.1
